% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_api_keys.R
\name{binance_api}
\alias{binance_api}
\title{Set the binance API Key and secret}
\usage{
binance_api(bin_key, bin_secret)
}
\arguments{
\item{bin_key}{A character string with your Binance API Key.}

\item{bin_secret}{A character string with your Binance API Secret.}
}
\value{
Invisibly returns API key and Secret once set. Use print method to view.
}
\description{
Set the binance API Key and secret
}
\details{
The Binance API key and Secret must be set prior to load crypto timeseries that needs it.
You can obtain an API key at your binance account (https://www.binance.com/en/support/faq/how-to-create-an-api-key-for-a-lead-trading-portfolio-2bec848b904b422197ce121d0925f20b).
}
\examples{
\dontrun{
binance_api("YOUR_API_KEY")
crypto_test = load_crypto_timeseries(pair = "ETHUSDT", interval="1m", start_date="2021-01-01", end_date="2021-07-01")
}

}
