% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invest.data-package.R
\docType{package}
\name{invest.data-package}
\alias{invest.data}
\alias{invest.data-package}
\title{invest.data: Fetch financial data from online sources}
\description{
This package provides functions to fetch financial data from online sources. It is intended to be used by other domestic projects.
}
\author{
\strong{Maintainer}: Henrique Ramos \email{first.last@example.com}

}
\keyword{internal}
