% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_timeseries.R
\name{load_yahoo_dividends}
\alias{load_yahoo_dividends}
\title{load_yahoo_dividends ############}
\usage{
load_yahoo_dividends(symbols, start_date, end_date = as.character(Sys.Date()))
}
\arguments{
\item{symbols}{Character vector. Stock ticker(s) to retrieve dividends for.}

\item{start_date}{Character. Start date in "YYYY-MM-DD".}

\item{end_date}{Character. End date in "YYYY-MM-DD". Default is today's date.}
}
\value{
A list with:
\describe{
\item{data}{Tibble with dividends data.}
\item{errors}{Tibble with symbols that could not be retrieved, including error message.}
}
}
\description{
Load dividends from Yahoo Finance using tidyquant.
}
\examples{
dividends <- load_yahoo_dividends(
  symbols = c("AAPL", "MSFT"),
  start_date = "2020-01-01",
  end_date   = "2022-01-01"
)
}
