% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_api_keys.R
\name{set_tiingo_api_key}
\alias{set_tiingo_api_key}
\title{Set or Retrieve the Tiingo API Key}
\usage{
set_tiingo_api_key(tiingo_key = NULL, overwrite = FALSE)
}
\arguments{
\item{tiingo_key}{Optional. A character string with your Tiingo API Key.
If not provided, the function will first look in \code{options("tiingo_key")},
then in the system environment variable \code{TIINGO_KEY},
and finally attempt to retrieve it from the keyring.}

\item{overwrite}{Logical, default = FALSE. Whether to overwrite the
key already set in options().}
}
\value{
Invisibly returns the Tiingo API key once set.
}
\description{
Set or Retrieve the Tiingo API Key
}
\details{
The Tiingo API key must be set before calling functions that require it.

Order of precedence for retrieving the key:
\enumerate{
\item Direct input via \code{tiingo_key}
\item Already set in \code{options("tiingo_key")}
\item Environment variable \code{Sys.getenv("TIINGO_KEY")}
\item Securely stored in the system keyring under the name \code{"tiingo_key"}
}
}
\examples{
\dontrun{
# Direct input
set_tiingo_api_key("YOUR_API_KEY")

# Or if previously stored with:
# keyring::key_set("tiingo_key")
set_tiingo_api_key()
}

}
