% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_api_keys.R
\name{tiingo_api}
\alias{tiingo_api}
\title{Set the tiingo API Key}
\usage{
tiingo_api(tiingo_key)
}
\arguments{
\item{tiingo_key}{A character string with your Tiingo API Key.}
}
\value{
Invisibly returns API key once set. Use print method to view.
}
\description{
Set the tiingo API Key
}
\details{
The Tiingo
API key must be set prior to load stock timeseries that needs it.
You can obtain an API key at your Tiingo account (https://api.tiingo.com/).
}
\examples{
\dontrun{
tiingo_api("YOUR_API_KEY")
stock_test = load_stock_timeseries(symbol = "AAPL", start_date="2021-01-01", end_date="2021-07-01")
}

}
