% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_timeseries.R
\name{load_crypto_timeseries}
\alias{load_crypto_timeseries}
\title{load_crypto_timeseries ############}
\usage{
load_crypto_timeseries(
  pair,
  interval,
  chunk_size = 1000,
  start_date,
  end_date,
  source = "binance"
)
}
\arguments{
\item{pair}{Character vector. Crypto pair(s), e.g. "ETHUSDT".}

\item{interval}{Character. One of '1m','3m','5m','15m','30m','1h',
'2h','4h','6h','8h','12h','1d','3d','1w','1M'.}

\item{chunk_size}{Integer. Number of observations per query (max 1000).}

\item{start_date}{Character in "YYYY-MM-DD".}

\item{end_date}{Character in "YYYY-MM-DD".}

\item{source}{Character. Currently only "binance" supported.}
}
\value{
A list with:
\describe{
\item{data}{Tibble with OHLCV timeseries.}
\item{errors}{Tibble with failed query ranges or NULL.}
}
}
\description{
Load crypto timeseries from Binance.
}
\examples{
# Load 1 minute time series for ETHUSDT
crypto_test <- load_crypto_timeseries(
  pair = "ETHUSDT",
  interval = "1m",
  start_date = "2021-01-01",
  end_date   = "2021-01-02"
)
}
