% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_timeseries.R
\name{load_stock_timeseries}
\alias{load_stock_timeseries}
\title{load_stock_timeseries ############}
\usage{
load_stock_timeseries(
  symbol,
  interval,
  chunk_size = NULL,
  start_date,
  end_date,
  adjust_splits = TRUE
)
}
\arguments{
\item{symbol}{Character vector. Stock ticker(s).}

\item{interval}{Character. One of '1m','3m','5m','15m','30m',
'1h','2h','4h','6h','8h','12h','1d'.}

\item{chunk_size}{Integer. Number of observations per query (max 1000).}

\item{start_date}{Character in "YYYY-MM-DD".}

\item{end_date}{Character in "YYYY-MM-DD".}

\item{adjust_splits}{Logical. Adjust prices for splits (Tiingo only).}
}
\value{
A list with:
\describe{
\item{data}{Tibble with OHLCV timeseries.}
\item{errors}{Tibble with failed query ranges or NULL.}
}
}
\description{
Load stock timeseries from Yahoo Finance (daily) or Tiingo (intraday).
}
\examples{
# Load daily time series for AAPL from 2021-10-01 to 2022-02-01
stock_test_day = load_stock_timeseries(symbol = "AAPL",
interval="1d",
start_date="2021-10-01",
end_date="2022-02-01")
# Load daily time series for AAPL and MSFT from 2021-10-01 to 2022-02-01
stock_test_multi = load_stock_timeseries(symbol = c("AAPL", "MSFT"),
interval="1d",
start_date="2021-10-01",
end_date="2022-02-01")
# # Load hourly time series for AAPL from 2021-10-01 to 2022-02-01
stock_test_hour = load_stock_timeseries(symbol = "AAPL",
interval="1h",
start_date="2021-10-01",
end_date="2022-02-01")

}
