% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_timeseries.R
\name{load_yahoo_dividends}
\alias{load_yahoo_dividends}
\title{load_yahoo_dividends
Load dividends from yahoo finance using tidyquant.}
\usage{
load_yahoo_dividends(
  list_tikers,
  start_date,
  end_date = as.character(Sys.Date())
)
}
\arguments{
\item{list_tikers}{A vector of stock symbols}

\item{start_date}{Start date of the time series (YYYY-MM-DD)}

\item{end_date}{End date of the time series (YYYY-MM-DD). Default is today's date.}
}
\value{
A tibble with the dividends data
}
\description{
load_yahoo_dividends
Load dividends from yahoo finance using tidyquant.
}
\examples{
#' # Load dividends for AAPL and MSFT from 2020-01-01 to 2022-01-01
dividends = load_yahoo_dividends(list_tikers = c("AAPL", "MSFT"), start_date = "2020-01-01", end_date = "2022-01-01")

}
